@@ -1,14 +1,14 @@
 source 'https://rubygems.org'
 git_source(:github) { |repo| "https://github.com/#{repo}.git" }
 
-ruby '2.5.1'
+ruby '2.6.3'
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '~> 5.2.2'
+gem 'rails', '~> 6.0'
 # Use postgresql as the database for Active Record
 gem 'pg', '>= 0.18', '< 2.0'
 # Use Puma as the app server
-gem 'puma', '~> 3.12'
+gem 'puma', '~> 4.1.0'
 
 # Reduces boot times through caching; required in config/boot.rb
 gem 'bootsnap', '>= 1.1.0', require: false
@@ -16,6 +16,7 @@ gem 'bootsnap', '>= 1.1.0', require: false
 group :development, :test do
   # Call 'byebug' anywhere in the code to stop execution and get a debugger console
   gem 'byebug', platforms: [:mri, :mingw, :x64_mingw]
+  gem 'listen'
 end
 
 # Windows does not include zoneinfo files, so bundle the tzinfo-data gem
@@ -27,7 +28,7 @@ gem 'devise_token_auth' # API Authentication
 
 gem 'has_scope' # Scope generation
 
-gem 'counter_culture', '~> 1.9.2' # Counter cache
+gem 'counter_culture' # Counter cache
 
 gem 'redis' # Caching and actioncable
 
@@ -38,3 +39,5 @@ gem 'pg_search' # Text search for rails models
 gem 'sidekiq' # Async background task management
 
 gem 'by_star' # Relative date/time related queries
+
+gem 'kaminari' # Pagination
